Rename module to `library`
