+++ core/src/lib.rs
@@ -8,7 +8,7 @@ pub mod event;
 pub mod job;
 pub mod markdown;
 // pub mod storage;
-pub mod storage2;
+pub mod library;
 pub mod tool;
 pub mod vector_store;
